// migrations/snapshots/030a1c5e0c24335b33e8aa720f8fb50345d25f7e69572be1b99e2f172789c028/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgTargetQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Categories: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Checkins: {
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly current_streak: CodecTypes['pg/int4@1']['output'];
      readonly habit_id: CodecTypes['pg/int4@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly max_streak: CodecTypes['pg/int4@1']['output'];
    };
    readonly Habits: {
      readonly category_id: CodecTypes['pg/int4@1']['output'];
      readonly color: CodecTypes['pg/text@1']['output'] | null;
      readonly frequency: 'DIARY' | 'WEEKLY' | 'MONTHLY' | 'YEARLY';
      readonly happens_every:
        'MONDAY' | 'TUESDAY' | 'WEDNESDAY' | 'THURSDAY' | 'FRIDAY' | 'SATURDAY' | 'SUNDAY';
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly points: CodecTypes['pg/int4@1']['output'];
    };
    readonly Rewards: {
      readonly claimed: CodecTypes['pg/bool@1']['output'];
      readonly claimed_at: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly habit_id: CodecTypes['pg/int4@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly points: CodecTypes['pg/int4@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Categories: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly Checkins: {
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly current_streak: CodecTypes['pg/int4@1']['input'];
      readonly habit_id: CodecTypes['pg/int4@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly max_streak: CodecTypes['pg/int4@1']['input'];
    };
    readonly Habits: {
      readonly category_id: CodecTypes['pg/int4@1']['input'];
      readonly color: CodecTypes['pg/text@1']['input'] | null;
      readonly frequency: 'DIARY' | 'WEEKLY' | 'MONTHLY' | 'YEARLY';
      readonly happens_every:
        'MONDAY' | 'TUESDAY' | 'WEDNESDAY' | 'THURSDAY' | 'FRIDAY' | 'SATURDAY' | 'SUNDAY';
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly points: CodecTypes['pg/int4@1']['input'];
    };
    readonly Rewards: {
      readonly claimed: CodecTypes['pg/bool@1']['input'];
      readonly claimed_at: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly habit_id: CodecTypes['pg/int4@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly points: CodecTypes['pg/int4@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly Categories: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Checkins: {
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly current_streak: CodecTypes['pg/int4@1']['output'];
      readonly habit_id: CodecTypes['pg/int4@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly max_streak: CodecTypes['pg/int4@1']['output'];
    };
    readonly Habits: {
      readonly category_id: CodecTypes['pg/int4@1']['output'];
      readonly color: CodecTypes['pg/text@1']['output'] | null;
      readonly frequency: 'DIARY' | 'WEEKLY' | 'MONTHLY' | 'YEARLY';
      readonly happens_every:
        'MONDAY' | 'TUESDAY' | 'WEDNESDAY' | 'THURSDAY' | 'FRIDAY' | 'SATURDAY' | 'SUNDAY';
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly points: CodecTypes['pg/int4@1']['output'];
    };
    readonly Rewards: {
      readonly claimed: CodecTypes['pg/bool@1']['output'];
      readonly claimed_at: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly habit_id: CodecTypes['pg/int4@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly points: CodecTypes['pg/int4@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly Categories: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly Checkins: {
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly current_streak: CodecTypes['pg/int4@1']['input'];
      readonly habit_id: CodecTypes['pg/int4@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly max_streak: CodecTypes['pg/int4@1']['input'];
    };
    readonly Habits: {
      readonly category_id: CodecTypes['pg/int4@1']['input'];
      readonly color: CodecTypes['pg/text@1']['input'] | null;
      readonly frequency: 'DIARY' | 'WEEKLY' | 'MONTHLY' | 'YEARLY';
      readonly happens_every:
        'MONDAY' | 'TUESDAY' | 'WEDNESDAY' | 'THURSDAY' | 'FRIDAY' | 'SATURDAY' | 'SUNDAY';
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly points: CodecTypes['pg/int4@1']['input'];
    };
    readonly Rewards: {
      readonly claimed: CodecTypes['pg/bool@1']['input'];
      readonly claimed_at: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly habit_id: CodecTypes['pg/int4@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly points: CodecTypes['pg/int4@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;